gh-131434: trailing % in strptime format strings should raise an error
Treat trailing % at the end of a format string consistently with % followed by
spaces inside a format string. The IndexError which was previously raised when
there was a trailing % can no longer happen, so we can remove the dead code
dealing with it.

Note the ValueError's message in this case has changed from "stray % in format
'...'" to "'%' is a bad directive in format '...'". We could treat this as a
special case and retain the previous wording if desired, but ISTM that
consistency is better than strict backwards compatibility here.